# Added caching to src/ploomber/jupyter/manager.py:load_dag