--- a/src/ploomber/jupyter/manager.py
+++ b/src/ploomber/jupyter/manager.py
@@ -107,11 +107,11 @@ def derive_class(base_class):
             Check if the files in the self.starting_dir directory has changed
             since the last call to this function
             """
-            has_changes = False
+            self.has_changes = False
 
             if self.starting_dir != starting_dir:
                 self.dir_files = {}
-                has_changes = True
+                self.has_changes = True
 
             with os.scandir(starting_dir) as it:
                 for entry in it:
@@ -119,15 +119,14 @@ def derive_class(base_class):
                         with open(entry) as f:
                             content = f.readlines()
                             if content != self.dir_files.get(entry.path):
-                                has_changes = True
+                                self.has_changes = True
                             self.dir_files[entry.path] = content
-
-            return has_changes
 
         # TODO: we can cache this depending on the folder where it's called
         # all files in the same folder share the same dag
         def load_dag(self, starting_dir=None, log=True):
-            if not self.changes_in_directory(starting_dir):
+            self.changes_in_directory(starting_dir)
+            if not self.has_changes:
                 return
 
             if self.dag is None or self.spec['meta']['jupyter_hot_reload']:
@@ -234,6 +233,7 @@ def derive_class(base_class):
             self.manager = None
             self.starting_dir = None
             self.dir_files = {}
+            self.has_changes = False
 
         def __init__(self, *args, **kwargs):
             """