Resolved issue #382

I don't really know how I can test this, so for now there is no new test added for this feature.
Below are some benchmarking before and after the changes (time went down from 93ms to 1ms when the directory is already loaded)

Before the changes:
```py
Python 3.9.7 | packaged by conda-forge | (default, Sep 29 2021, 19:23:11)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.28.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: from ploomber.jupyter.manager import derive_class
   ...: from jupytext.contentsmanager import TextFileContentsManager
   ...: %cd ~/ploomber/tests/assets/nbs
   ...: PCM = derive_class(TextFileContentsManager)
   ...: cm = PCM()
~/ploomber/tests/assets/nbs

In [2]: %time cm.load_dag()
~/ploomber/src/ploomber/spec/dagspec.py:336: UserWarning: The following placeholders are declared in t
he environment but unused in the spec: {'sample'}
  warnings.warn('The following placeholders are declared in the '
CPU times: user 178 ms, sys: 16.2 ms, total: 194 ms
Wall time: 198 ms

In [3]: %time cm.load_dag()
~/ploomber/src/ploomber/spec/dagspec.py:336: UserWarning: The following placeholders are declared in t
he environment but unused in the spec: {'sample'}
  warnings.warn('The following placeholders are declared in the '
CPU times: user 83.9 ms, sys: 4.7 ms, total: 88.6 ms
Wall time: 92.5 ms
```
After the changes: 
```py
In [1]: from ploomber.jupyter.manager import derive_class
   ...: from jupytext.contentsmanager import TextFileContentsManager
   ...: %cd ~/ploomber/tests/assets/nbs
   ...: PCM = derive_class(TextFileContentsManager)
   ...: cm = PCM()
~/ploomber/tests/assets/nbs

In [2]: %time cm.load_dag()
~/ploomber/src/ploomber/spec/dagspec.py:336: UserWarning: The following placeholders are declared in t
he environment but unused in the spec: {'sample'}
  warnings.warn('The following placeholders are declared in the '
CPU times: user 171 ms, sys: 13.2 ms, total: 184 ms
Wall time: 187 ms

In [3]: %time cm.load_dag()
CPU times: user 399 µs, sys: 413 µs, total: 812 µs
Wall time: 1.03 ms
```